requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
environ==1.0
idna==2.8
influxdb==5.2.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
paho-mqtt==1.5.0
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2019.3
requests==2.22.0
six==1.13.0
typed-ast==1.4.1
urllib3==1.25.7
wrapt==1.11.2
